Shift nots tests into their own test file.
